Fix bullhead Marshmallow build
Copying files to system/vendor isn't allowed when you are building a
vendor image.  Skip the pittpatt copy statements when the device is
bullhead (since we build a vendor image).

Rather than checking the device here, what we should really be doing
is skipping those copy statements for any time a vendor image is being
built.  Inside the "build" project, the statement "ifdef
BOARD_VENDORIMAGE_FILE_SYSTEM_TYPE" is used to do that, but that
doesn't work here because of the way the opengapps makefiles is
included (they are processed before
device/lge/bullhead/BoardConfig.mk).